+++ src/Symfony/Component/DependencyInjection/ContainerAwareTrait.php
@@ -11,10 +11,14 @@
 
 namespace Symfony\Component\DependencyInjection;
 
+trigger_deprecation('symfony/dependency-injection', '6.4', '"%s" is deprecated, use dependency injection instead.', ContainerAwareTrait::class);
+
 /**
  * ContainerAware trait.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @deprecated since Symfony 6.4, use dependency injection instead
  */
 trait ContainerAwareTrait
 {